requirements.txt
django==2.2
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt

